requirements.txt
fastmcp==2.2.8
aiomysql>=0.2.0
python-dotenv>=1.1.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.23.0